+++ src/main/java/com/example/paymentsvc/model/response/MerchantResponse.java
@@ -3,12 +3,14 @@ package com.example.paymentsvc.model.response;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class MerchantResponse {
 
     private Long id;
     private String name;
-    private Integer commission;
+    private BigDecimal commission;
     private Long categoryId;
 }
